* Make an example for capturetex
